fix compilation error

+++ clang_delta/ReplaceOneLevelTypedefType.cpp
@@ -79,7 +79,7 @@ void ReplaceOneLevelTypedefType::HandleTranslationUnit(ASTContext &Ctx)
 
   Ctx.getDiagnostics().setSuppressAllDiagnostics(false);
 
-  if (TransformationCounter <= ValidTypedefTypes.size()) {
+  if (TransformationCounter <= (unsigned)ValidTypedefTypes.size()) {
     auto TL = ValidTypedefTypes[TransformationCounter - 1];
     rewriteTypedefType(TL, TL.getTypedefNameDecl());
     removeTypedefs(TL.getTypedefNameDecl());
